# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CarFax Web API
  description: API для веб-сервиса отчётов по VIN
  version: 1.0.0
  contact:
    name: CarFax Web Team
    email: support@carfax-web.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Локальная разработка
  - url: https://api.carfax-web.com/api/v1
    description: Продакшн сервер
security:
  - BearerAuth: []
paths:
...
  /reports/request:
    post:
      tags:
        - Отчёты
      summary: Запрос отчёта по VIN
      description: Создание запроса на генерацию отчёта по VIN номеру
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestReportRequest'
            example:
              vin: "1HGBH41JXMN109186"
      responses:
        '201':
          description: Запрос отчёта создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestReportResponse'
  /reports/{id}:
    get:
      tags:
        - Отчёты
      summary: Получение информации об отчёте
      description: Возвращает детальную информацию об отчёте
      parameters:
        - name: id
          in: path
          required: true
          description: ID отчёта
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация об отчёте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '404':
          description: Отчёт не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reports/{id}/download:
    get:
      tags:
        - Отчёты
      summary: Скачивание отчёта
      description: Получение ссылки для скачивания PDF отчёта
      parameters:
        - name: id
          in: path
          required: true
          description: ID отчёта
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: URL для скачивания
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloadUrl:
                    type: string
                    format: uri
                    description: Ссылка для скачивания PDF
        '404':
          description: PDF файл не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RequestReportRequest:
      type: object
      required:
        - vin
      properties:
        vin:
          type: string
          pattern: '^[A-HJ-NPR-Z0-9]{17}$'
          description: VIN номер автомобиля (17 символов)
    RequestReportResponse:
      type: object
      properties:
        reportId:
          type: string
          format: uuid
          description: ID созданного отчёта
        status:
          type: string
          enum: [available, payment_required, generating]
          description: Статус отчёта
        downloadUrl:
          type: string
          format: uri
          description: Ссылка для скачивания (если доступно)
        price:
          type: number
          format: float
          description: Цена отчёта (если требуется оплата)
        message:
          type: string
          description: Сообщение о статусе
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vin:
          type: string
        user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, generating, completed, error]
        price:
          type: number
          format: float
        pdf_file_name:
          type: string
        generated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        downloadUrl:
          type: string
          format: uri
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP статус код
        message:
          type: string
          description: Сообщение об ошибке
        error:
          type: string
          description: Тип ошибки